login and sign up functionality implemented
