+++ src/App.js
@@ -1,6 +1,6 @@
-import React from "react";
 import "./App.css";
 import { Route, Routes } from "react-router-dom";
+import ProtectedRoutes from "./Router/Protected";
 import Home from "./pages/Home/Home";
 import About from "../src/Components/Aboutpage/About";
 import Explore from "./pages/Explore/Explore";
@@ -8,7 +8,7 @@ import Blog from "./pages/Blog/Blog";
 import Contact from "./pages/Contact/Contact";
 import SignIn from "../src/Components/SignIn/SignIn";
 import SignUp from "../src/Components/SignUp/SignUp";
-import Register from "./pages/Register/RegisterTest";
+import Register from "./pages/Register/Register";
 import EventRegister from "./pages/Register/EventRegister";
 import CreateProfile from "./pages/CreateProfile/CreateProfile";
 import SpeakerProfile from "./pages/SpeakerProfile/SpeakerProfile";
@@ -21,6 +21,11 @@ import Chat from "./pages/Chat/Chat";
 import Dashboard from "./pages/Dashboard/Dashboard";
 
 function App() {
+  function isLoggedIn() {
+    const isAuthenticated = !!sessionStorage.getItem("token");
+    return isAuthenticated;
+  }
+
   return (
     <Routes>
       <Route path="/" element={<Home />} />
@@ -32,7 +37,14 @@ function App() {
       <Route path="/signup" element={<SignUp />} />
       <Route path="/register" element={<Register />} />
       <Route path="/event-register" element={<EventRegister />} />
-      <Route path="/create-profile" element={<CreateProfile />} />
+      <Route
+        path="/create-profile"
+        element={
+          // <ProtectedRoutes isLoggedIn={isLoggedIn()}>
+            <CreateProfile />
+          // </ProtectedRoutes>
+        }
+      />
       <Route path="/speaker-profile" element={<SpeakerProfile />} />
       <Route path="/book-me" element={<Bookme />} />
       <Route path="/chat" element={<Chat />} />
@@ -40,7 +52,14 @@ function App() {
       <Route path="/oops" element={<Oops />} />
       <Route path="/congratulation" element={<Congratulation />} />
       <Route path="/congratulation2" element={<Congratulation2 />} />
-      <Route path="/dashboard" element={<Dashboard />} />
+      <Route
+        path="/dashboard"
+        element={
+          // <ProtectedRoutes isLoggedIn={isLoggedIn()}>
+            <Dashboard />
+          // </ProtectedRoutes>
+        }
+      />
     </Routes>
   );
 }
