+++ src/pages/Register/Register.js
@@ -5,85 +5,16 @@ import Logo from "../../images/footerLogo.png";
 import { FcGoogle } from "react-icons/fc";
 import { FaFacebookF, FaTwitter } from "react-icons/fa";
 import toast from "react-hot-toast";
-import { useFormik } from "formik";
 import { useAuth } from "../../contexts/AuthContext";
+
+import useInput from "./useInput";
+import styles from "./Register.module.css";
 
 function Register() {
   const navigate = useNavigate();
 
-  const { signup, googleSignIn } = useAuth();
+  const { googleSignIn } = useAuth();
   const [loading, setLoading] = useState(false);
-
-  //form submit function
-  const onSubmit = async (values) => {
-    setLoading(true);
-    await signup(values.email, values.password)
-      .then((userCredential) => {
-        const user = userCredential.user;
-        console.log(user);
-        setLoading(false);
-        toast.success(
-          "Registration Successful, Please complete your registration",
-          {
-            duration: 4000,
-            position: "top-center",
-
-            // Styling
-            style: { fontSize: "13px" },
-            className: "",
-          }
-        );
-        navigate("/create-profile");
-      })
-      .catch((error) => {
-        const errorCode = error.code;
-        toast.error(errorCode);
-        setLoading(false);
-      });
-  };
-
-  //form validation
-  const validate = (values) => {
-    const errors = {};
-
-    if (!values.password) {
-      errors.password = "Required";
-    } else if (values.password.length < 6) {
-      errors.password = "Must be more than 6 characters";
-    } else if (values.password.includes(" ")) {
-      errors.password = "Invalid Password";
-    }
-
-    if (!values.confirmPassword) {
-      errors.confirmPassword = "Required";
-    } else if (values.password.length < 6) {
-      errors.confirmPassword = "Must be more than 6 characters";
-    } else if (values.password !== values.confirmPassword) {
-      errors.confirmPassword = "Confirm password is incorrect";
-    } else if (values.confirmPassword.includes(" ")) {
-      errors.confirmPassword = "Invalid Password";
-    }
-
-    if (!values.email) {
-      errors.email = "Required";
-    } else if (
-      !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.email)
-    ) {
-      errors.email = "Invalid email address";
-    }
-
-    return errors;
-  };
-
-  const formik = useFormik({
-    initialValues: {
-      email: "",
-      password: "",
-      confirmPassword: "",
-    },
-    validate,
-    onSubmit,
-  });
 
   const handleGoogleSignUp = async () => {
     try {
@@ -104,61 +35,144 @@ function Register() {
       return null;
     }
   };
+  const {
+    value: emailInputvalue,
+    isValid: enteredEmailIsValid,
+    hasError: emailInputHasError,
+    valueChangeHandler: emailChangeHandler,
+    inputBlurHandler: emailBlurHandler,
+    reset: resetEmailInput,
+  } = useInput((value) => value.includes("@"));
+
+  const {
+    value: passwordInputValue,
+    isValid: enteredPasswordIsValid,
+    hasError: passwordInputHasError,
+    valueChangeHandler: passwordChangeHandler,
+    inputBlurHandler: passwordBlurHandler,
+    reset: resetPasswordInput,
+  } = useInput((value) => value.length >= 6 && value.trim() !== "");
+
+  const {
+    value: confirmPasswordInputValue,
+    isValid: enteredconfirmPasswordIsValid,
+    hasError: confirmPasswordInputHasError,
+    valueChangeHandler: confirmPasswordChangeHandler,
+    inputBlurHandler: confirmPasswordBlurHandler,
+    reset: resetConfirmPasswordInput,
+  } = useInput(
+    (value) =>
+      value.trim() !== "" &&
+      value.length >= 6 &&
+      value.trim().match(passwordInputValue)
+  );
+
+  let formIsValid = false;
+
+  if (
+    emailInputvalue &&
+    passwordInputValue &&
+    confirmPasswordInputValue === passwordInputValue
+  ) {
+    formIsValid = true;
+  }
+
+  const signUpSubmitHandler = async (e) => {
+    try {
+      e.preventDefault();
+
+      if (!formIsValid) {
+        return;
+      }
+
+      if (
+        !emailInputvalue &&
+        !passwordInputValue &&
+        !confirmPasswordInputValue
+      ) {
+        return;
+      }
+
+      const saveUserData = await fetch(
+        "https://spefind-server.onrender.com/api/auth/speaker/signup",
+        {
+          method: "POST",
+          body: JSON.stringify({
+            email: emailInputvalue,
+            password: passwordInputValue,
+          }),
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+      console.log(saveUserData);
+      setLoading(true);
+
+      const data = await saveUserData.json();
+      console.log(data);
+
+      if (saveUserData.ok) {
+        setLoading(false);
+        toast.success(`${data.message}, " Please complete your registration"`, {
+          duration: 4000,
+          position: "top-center",
+
+          // Styling
+          style: { fontSize: "13px" },
+          className: "",
+        });
+        navigate("/create-profile");
+      }
+
+      if (!saveUserData.ok) {
+        setLoading(false);
+        toast.error(`${data.message},`, {
+          duration: 4000,
+          position: "top-center",
+
+          // Styling
+          style: { fontSize: "13px" },
+          className: "",
+        });
+        throw new Error();
+      }
+      if (!saveUserData.ok) {
+        setLoading(false);
+        throw new Error();
+      }
+      resetEmailInput();
+      resetPasswordInput();
+      resetConfirmPasswordInput();
+    } catch (error) {
+      console.log(error);
+    }
+  };
 
   return (
     <div>
       <div className="Bg-img">
         <img src={Logo} alt="Logo" className="regLogo" />
-        <form className="JoinForm" onSubmit={formik.handleSubmit}>
+        <form className="JoinForm" onSubmit={signUpSubmitHandler}>
           <div className="ovalBG"></div>
-          {/* <div className="NameCard">
-                    <div className="me-3 mt-1">
-                        <label className="labell">First Name</label><br/>
-                        <input 
-                          type="text" 
-                          placeholder="First Name"  
-                          className="regInput1" 
-                          name='firstName'
-                          {...formik.getFieldProps('firstName')}
-                          required
-                        />
-                    { formik.errors.firstName ? <div className='text-danger mb-3 ' style={{fontSize: '0.7rem', textAlign: 'left', fontWeight: 'bold'}}>{ formik.errors.firstName }</div> : null }
-                    </div>
-
-                    <div className="mt-1">
-                        <label className="labell">Last Name</label><br/>
-                        <input 
-                        type="text" 
-                        placeholder="Last Name"  
-                        className="regInput1" 
-                        name='lastName'
-                          {...formik.getFieldProps('lastName')}
-                        required
-                        />
-                    { formik.errors.lastName ? <div className='text-danger mb-3 ' style={{fontSize: '0.7rem', textAlign: 'left', fontWeight: 'bold'}}>{ formik.errors.lastName }</div> : null }
-                    </div>
-                </div> */}
           <div>
             <label className="labelForm">E-mail Address</label>
             <input
               type="email"
               name="email"
-              {...formik.getFieldProps("email")}
-              className="regInput mb-0"
+              className={
+                emailInputHasError
+                  ? `${styles.invalidInput} regInput mb-0`
+                  : "regInput mb-0"
+              }
               placeholder="Email address"
+              value={emailInputvalue}
+              onChange={emailChangeHandler}
+              onBlur={emailBlurHandler}
             />
-            {formik.errors.email ? (
-              <div
-                className="text-danger mb-1"
-                style={{
-                  fontSize: "0.7rem",
-                  textAlign: "left",
-                  fontWeight: "bold",
-                }}
-              >
-                {formik.errors.email}
-              </div>
-            ) : null}
+            {emailInputHasError && !enteredEmailIsValid && (
+              <p className={styles.errorText}>Please enter a valid email !</p>
+            )}
           </div>
 
           <div>
@@ -167,23 +181,22 @@ function Register() {
               type="password"
               label="Create password"
               name="password"
-              {...formik.getFieldProps("password")}
               required
-              className="regInput"
+              className={
+                passwordInputHasError
+                  ? `${styles.invalidInput} regInput mb-0`
+                  : "regInput mb-0"
+              }
               placeholder="Password"
+              value={passwordInputValue}
+              onChange={passwordChangeHandler}
+              onBlur={passwordBlurHandler}
             />
-            {formik.errors.password ? (
-              <div
-                className="text-danger mb-1"
-                style={{
-                  fontSize: "0.7rem",
-                  textAlign: "left",
-                  fontWeight: "bold",
-                }}
-              >
-                {formik.errors.password}
-              </div>
-            ) : null}
+            {passwordInputHasError && !enteredPasswordIsValid && (
+              <p className={styles.errorText}>
+                Password must be greater than 6 characters !
+              </p>
+            )}
           </div>
 
           <div>
@@ -191,25 +204,26 @@ function Register() {
             <input
               type="password"
               placeholder="Confirm password"
-              className="regInput"
+              className={
+                confirmPasswordInputHasError
+                  ? `${styles.invalidInput} regInput mb-0`
+                  : "regInput mb-0"
+              }
               name="confirmPassword"
-              {...formik.getFieldProps("confirmPassword")}
+              value={confirmPasswordInputValue}
+              onChange={confirmPasswordChangeHandler}
+              onBlur={confirmPasswordBlurHandler}
             />
-            {formik.errors.confirmPassword ? (
-              <div
-                className="text-danger mb-1"
-                style={{
-                  fontSize: "0.7rem",
-                  textAlign: "left",
-                  fontWeight: "bold",
-                }}
-              >
-                {formik.errors.confirmPassword}
-              </div>
-            ) : null}
+            {confirmPasswordInputHasError && !enteredconfirmPasswordIsValid && (
+              <p className={styles.errorText}>Passwords do not match !</p>
+            )}
           </div>
 
-          <button className=" btn btnSign-up" type="submit">
+          <button
+            className=" btn btnSign-up"
+            type="submit"
+            disabled={!formIsValid}
+          >
             {loading ? "Signing up..." : "Sign up"}
           </button>
           <p className="or">OR</p>
